+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 ### 1.7.1 (Next)
 - Fix Packer generated data interpolation for `env_vars` and `test_files` parameters.
 - Minor improvement to warning logs.
+- Block on `local` `install_cmd` before initiating local Testinfra test execution.
 
 ### 1.7.0
 - Add `retries` parameter.